Registration refactor
