+++ app/src/main/java/com/dj/insulink/auth/ui/ForgotPasswordScreen.kt
@@ -1,4 +1,4 @@
-package com.dj.insulink.auth.ui.screen
+package com.dj.insulink.auth.ui
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -31,7 +31,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
